requirements.txt
Django
dj-database-url
dj-static
gunicorn
Unipath
python-decouple
bleach
whitenoise==3.2
djangorestframework
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt
